Add pycore_crossinterp.h.
