+++ Modules/_xxsubinterpretersmodule.c
@@ -6,6 +6,7 @@
 #endif
 
 #include "Python.h"
+#include "pycore_crossinterp.h"   // struct _xid
 #include "pycore_initconfig.h"    // _PyErr_SetFromPyStatus()
 #include "pycore_modsupport.h"    // _PyArg_BadArgument()
 #include "pycore_pyerrors.h"      // _PyErr_ChainExceptions1()
